Refactored BaseProjectTestCase
