+++ tests/com/magento/idea/magento2plugin/BaseProjectTestCase.java
@@ -4,14 +4,11 @@
  */
 package com.magento.idea.magento2plugin;
 
-import com.intellij.codeInsight.daemon.impl.HighlightInfo;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.testFramework.fixtures.BasePlatformTestCase;
-import com.magento.idea.magento2plugin.bundles.InspectionBundle;
 import com.magento.idea.magento2plugin.indexes.IndexManager;
 import com.magento.idea.magento2plugin.project.Settings;
 import com.magento.idea.magento2plugin.magento.packages.File;
-import java.util.List;
 
 /**
  * Configure test environment with Magento 2 project
@@ -19,7 +16,6 @@ import java.util.List;
 abstract public class BaseProjectTestCase extends BasePlatformTestCase {
     private static final String testDataProjectPath = "testData" + File.separator + "project";
     private static final String testDataProjectDirectory = "magento2";
-    protected final InspectionBundle inspectionBundle = new InspectionBundle();
 
     @Override
     protected void setUp() throws Exception {
@@ -69,38 +65,4 @@ abstract public class BaseProjectTestCase extends BasePlatformTestCase {
         return StringUtil.trimEnd(getTestName(true), "Test");
     }
 
-    protected void assertHasHighlighting(String message) {
-        String highlightingNotFound = "Failed that documents contains highlighting with the description `%s`";
-
-        List<HighlightInfo> highlightingList = myFixture.doHighlighting();
-        if (highlightingList.isEmpty()) {
-            fail(String.format(highlightingNotFound, message));
-        }
-
-        for (HighlightInfo highlighting :
-                highlightingList) {
-            if (highlighting.getDescription() == null) continue;
-            if (highlighting.getDescription().equals(message)) {
-                return;
-            }
-        }
-        fail(String.format(highlightingNotFound, message));
-    }
-
-    protected void assertHasNoHighlighting(String message) {
-        String highlightingNotFound = "Failed that documents not contains highlighting with the description `%s`";
-
-        List<HighlightInfo> highlightingList = myFixture.doHighlighting();
-        if (highlightingList.isEmpty()) {
-            return;
-        }
-
-        for (HighlightInfo highlighting :
-                highlightingList) {
-            if (highlighting.getDescription() == null) continue;
-            if (highlighting.getDescription().equals(message)) {
-                fail(String.format(highlightingNotFound, message));
-            }
-        }
-    }
 }
