+++ tests/com/magento/idea/magento2plugin/inspections/graphqls/InspectionGraphqlsFixtureTestCase.java
@@ -5,10 +5,10 @@
 
 package com.magento.idea.magento2plugin.inspections.graphqls;
 
-import com.magento.idea.magento2plugin.BaseProjectTestCase;
+import com.magento.idea.magento2plugin.inspections.BaseInspectionsTestCase;
 import com.magento.idea.magento2plugin.magento.packages.File;
 
-abstract public class InspectionGraphqlsFixtureTestCase extends BaseProjectTestCase {
+abstract public class InspectionGraphqlsFixtureTestCase extends BaseInspectionsTestCase {
 
     private static final String testDataFolderPath = "testData" + File.separator + "inspections" + File.separator;
     private static final String fixturesFolderPath = "graphqls" + File.separator;
