+++ tests/com/magento/idea/magento2plugin/inspections/php/InspectionPhpFixtureTestCase.java
@@ -4,11 +4,10 @@
  */
 package com.magento.idea.magento2plugin.inspections.php;
 
-import com.magento.idea.magento2plugin.BaseProjectTestCase;
+import com.magento.idea.magento2plugin.inspections.BaseInspectionsTestCase;
 import com.magento.idea.magento2plugin.magento.packages.File;
-import java.util.List;
 
-abstract public class InspectionPhpFixtureTestCase extends BaseProjectTestCase {
+abstract public class InspectionPhpFixtureTestCase extends BaseInspectionsTestCase {
 
     private static final String testDataFolderPath = "testData" + File.separator + "inspections" + File.separator;
     private static final String fixturesFolderPath = "php" + File.separator;
